feat : [TSK-12] HCX-007 API로 업데이트
